support to monitor slaves in multiple masters.

+++ host/check_vm/check_vm.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 #
-# Copyright (C) 2014 VA Linux Systems Japan K.K.
-# Copyright (C) 2014 Fumihiko Kakuma <kakuma at valinux co jp>
+# Copyright (C) 2014,2015 VA Linux Systems Japan K.K.
+# Copyright (C) 2014,2015 Fumihiko Kakuma <kakuma at valinux co jp>
 # All Rights Reserved.
 #
 #    Licensed under the Apache License, Version 2.0 (the "License"); you may
@@ -26,22 +26,24 @@ MAX_CHECK=3
 SLEEP_SEC=15
 REMOVEOVL='n'
 RECREATE='y'
-CHECK_SLAVE='n'
-TARGETVMS=${TARGETVMS:-"slave1 slave2 slave3 slave4 slave5 slave6 jslave1"}
-JENKINS_URL=${JENKINS_URL:-http://master:8080/}
+CHECK_CNT=0
+CHECK_INT_JNK=0
+TARGETVMS=${TARGETVMS:-"master:slave1,slave2,slave3,slave4 jmaster:jslave1"}
+TARGET_SLAVES=""
 JENKINS_CLI=/usr/local/jenkins/jenkins-cli.jar
 OFFLINE_MAX=${OFFLINE_MAX:-600}
 declare -A SLAVE_STATUS
+declare -A SLAVE_MAT
 
 function usage {
 	cat <<- EOF
-	USAGE : $SCRIPT_NAME [-h -c max-check -i int -j -m max -r -s -t vm-name]
+	USAGE : $SCRIPT_NAME [-h -c max-check -i int -j check -m max -r -s -t vm-name]
 	        -h display this messages
 	        -c max existece check count
 	           default: 3
 	        -i interval to check vm
 	           default: 15
-	        -j check if slave is offline
+	        -j interval to check if slave is offline
 	        -m max loop times
 	           runs until killed if max is zero
 	           default: 4
@@ -49,18 +51,21 @@ function usage {
 	           default: save an old ovl image file as xxx.old
 	        -s only start vm
 	           default: recreate an ovl image file
-	        -t specify target vm
+	        -t specify target slaves and master pair
 	           default: $TARGETVMS
 	        and you can use the following environment vaiables
 	          IMAGES_DIR: a directory has image file
-	          TARGETVMS: specify target vm(same to -t option)
+	          TARGETVMS: specify target slaves and master pair(same to -t option)
 	EOF
 }
 
 function jenkins_cli() {
     local cmd=$1
-    local slave=$2
-    java -jar $JENKINS_CLI -s $JENKINS_URL $cmd $slave
+    local master=$2
+    local slave=$3
+    local url="http://$master:8080"
+    echo ">>> jenkins_cli argument: $slave $url"
+    java -jar $JENKINS_CLI -s $url $cmd $slave
 }
 
 function check_slave_status {
@@ -68,10 +73,12 @@ function check_slave_status {
     local cli_res
     local count
 
-    if [[ $CHECK_SLAVE == 'n' ]]; then
+    if [[ $CHECK_INT_JNK -eq 0 ]]; then
+        echo "Checking jenkins status is unavailable."
         return 0
     fi
-    cli_res=$(jenkins_cli get-node $slave)
+    sleep 1
+    cli_res=$(jenkins_cli get-node ${SLAVE_MAT[$slave]} $slave)
     if [[ $? -ne 0 ]]; then
         echo "failed to get jenkins node status!!"
         return 0
@@ -81,7 +88,7 @@ function check_slave_status {
             echo "Currently $slave is offline."
             count=${SLAVE_STATUS[$slave]}
             SLAVE_STATUS[$slave]=$((++count))
-            if [[ $(($count*$SLEEP_SEC)) -gt $OFFLINE_MAX ]]; then
+            if [[ $(($SLEEP_SEC*$CHECK_INT_JNK*$count)) -gt $OFFLINE_MAX ]]; then
                 SLAVE_STATUS[$slave]=0
                 return 1
             else
@@ -109,7 +116,7 @@ function vm_exists {
     return 1
 }
 
-while getopts "hc:i:jm:rst:" opt; do
+while getopts "hc:i:j:m:rst:" opt; do
     case $opt in
       h)
         usage
@@ -122,7 +129,7 @@ while getopts "hc:i:jm:rst:" opt; do
         SLEEP_SEC=$OPTARG
         ;;
       j)
-        CHECK_SLAVE='y'
+        CHECK_INT_JNK=$OPTARG
         ;;
       m)
         MAX_LOOP=$OPTARG
@@ -149,25 +156,44 @@ else
     COUNT=0
     #SLEEP_SEC=$((60/$MAX_LOOP))
 fi
-for vm in $TARGETVMS; do
-    SLAVE_STATUS[$vm]=0
+for vm_set in $TARGETVMS; do
+    master=`echo $vm_set | cut -d':' -f1`
+    slave_set=`echo $vm_set | cut -d':' -f2`
+    slaves=${slave_set//,/ }
+    TARGET_SLAVES="$slaves $TARGET_SLAVES"
+    for slave in $slaves; do
+        SLAVE_MAT[$slave]=$master
+        SLAVE_STATUS[$slave]=0
+    done
 done
-declare -p MAX_LOOP SLEEP_SEC COUNT RECREATE REMOVEOVL TARGETVMS OFFLINE_MAX CHECK_SLAVE
-declare -p SLAVE_STATUS
+declare -p MAX_LOOP SLEEP_SEC COUNT RECREATE REMOVEOVL TARGETVMS TARGET_SLAVES
+declare -p OFFLINE_MAX CHECK_INT_JNK
+declare -p SLAVE_STATUS SLAVE_MAT
 echo "sleep every $SLEEP_SEC sec"
 while [ $COUNT -lt $MAX_LOOP ]; do
     sleep $SLEEP_SEC
-    echo "start loop"
+    echo "########## start loop ##########"
     virsh list
     date
-    for vm in $TARGETVMS; do
+    CHECK_CNT=$((++CHECK_CNT))
+    declare -p CHECK_CNT
+    if [[ $CHECK_INT_JNK -gt $CHECK_CNT ]]; then
+        echo "** SKIP checking jenkins status"
+    else
+        CHECK_CNT=0
+        declare -p CHECK_CNT
+    fi
+    for vm in $TARGET_SLAVES; do
         echo "check $vm"
         if vm_exists $vm; then
             echo "$vm exsits"
-            if ! check_slave_status $vm; then
-                echo "!!virsh destroy $vm"
-                virsh destroy $vm
+            if [[ $CHECK_CNT -eq 0 ]]; then
+                if ! check_slave_status $vm; then
+                    echo "!!virsh destroy $vm"
+                    virsh destroy $vm
+                fi
             fi
+            sleep 1
             continue
         fi
         virsh list
